examples/albef/requirements.txt
opencv-python==4.6.0.66
pytorch-lightning==1.6.0
